refactor(Display): update Display/components/Layer import/export path
