+++ webapp/app/containers/Display/components/Layer/index.tsx
@@ -19,3 +19,6 @@
  */
 
 export { default as LayerItem } from './LayerItem'
+export { default as LayerSimple } from './LayerSimple'
+export * from './util'
+export * from './types'
